datachanged

+++ website/views.py
@@ -1,6 +1,5 @@
 from os import stat
-from django.shortcuts import render
-from django.shortcuts import redirect
+from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.http import HttpResponse
 from django.conf import settings
@@ -11,6 +10,7 @@ import pyotp
 import random
 import json
 from django.urls import reverse
+from .models import User_Info
 import datetime
 
 from website.models import User_Info
@@ -25,20 +25,31 @@ def login(request):
 
         user_email_id = request.POST['user_email']
 
-        file = open('saving_mail_ids.txt', 'w')
-        file.write(user_email_id)
-        file.close()
+        existing_user_email_id = User_Info.objects.filter(user_email=user_email_id)
 
-        # stats for geenrating OTP
-        otp = hotp.now()
-        
-        msg_html = render_to_string('otp-email.html', {"otp": otp})
+        print(existing_user_email_id)
 
-        email = EmailMultiAlternatives(f'Fitlife.ai Account - {otp} is your OTP for secure access', '', settings.EMAIL_HOST_USER, [user_email_id])
-        email.attach_alternative(msg_html, "text/html")
-        email.send()
+        if len(existing_user_email_id) > 0:
+            
+            file = open('saving_mail_ids.txt', 'w')
+            file.write(user_email_id)
+            file.close()
 
-        return render(request,'login-next.html', {"user_email_id": user_email_id})
+            # stats for geenrating OTP
+            otp = hotp.now()
+            
+            msg_html = render_to_string('otp-email.html', {"otp": otp})
+
+            email = EmailMultiAlternatives(f'Fitlife.ai Account - {otp} is your OTP for secure access', '', settings.EMAIL_HOST_USER, [user_email_id])
+            email.attach_alternative(msg_html, "text/html")
+            email.send()
+
+            return render(request,'login-next.html', {"user_email_id": user_email_id})
+
+        else:
+            messages.error(request, "Please Register Yourself ! ")
+            return redirect('login')
+
 
     return render(request,'login.html')
 
@@ -54,12 +65,10 @@ def login_next(request):
         status = hotp.verify(str(user_otp))
 
         if status == True:
-            # messages.success(request, "Successfully Logged In")
-            # return HttpResponse('<h1>Successfully Logged In</h1>')
             return redirect('dashboard')
 
         else:
-            messages.warning(request, "Invalid OTP! Please try again")
+            messages.error(request, "Invalid OTP! Please try again")
             return redirect('login_next')
         
     return render(request,'login-next.html')
@@ -193,11 +202,12 @@ def personal_details(request):
 
 def body_details(request):
     if request.method == 'POST':
-        user_height = request.POST['height']
+        user_height_ft = int(request.POST['height'])
+        user_height_in = int(request.POST['height1'])
+        user_height = str(((user_height_in/12) + user_height_ft)*30.48)
         user_current_weight = request.POST['current-weight']
         user_targeted_weight = request.POST['targeted-weight']
-        print(user_height,user_current_weight,user_targeted_weight)
-
+    
         # updating the onboarding json file
         onboard_file = open('website/onboarding_stat.json', 'r+')
         onboard_data = json.load(onboard_file)
@@ -211,8 +221,8 @@ def body_details(request):
         #if male redirect to active_status_male
         if onboard_data['gender'] == 'male':
             return render(request, 'active-status-male.html', {'user_name': onboard_data['user_name']})
+        
         #else redirect to active_status_female
-
         else:
             return render(request, 'active-status-female.html', {'user_name': onboard_data['user_name']})
 
@@ -280,7 +290,7 @@ def main_goal(request):
                                         user_email = onboard_data['user_email'],
                                         user_age = onboard_data['user_age'],
                                         user_blood_group = onboard_data['user_blood_group'],
-                                        user_height = onboard_data['user_height'],
+                                        user_height  = onboard_data['user_height'],
                                         user_weight = onboard_data['user_current_weight'],
                                         user_activity_level = onboard_data['active_status'])
 
@@ -293,7 +303,8 @@ def main_goal(request):
             # user main goal not there in database
             # user targeted weight not there in database
 
-            return redirect('dashboard')
+           
+            return render(request,'dashboard.html',{'user_name': onboard_data['user_name']})
         
         else:
             return redirect('main_goal')
